requirements.txt
discord.py>=2.3.2
yt-dlp>=2024.3.10
pynacl>=1.5.0
